feat: map view android separation wip
