+++ android/src/main/kotlin/com/google/maps/flutter/navigation/GoogleMapsNavigationInspectorHandler.kt
@@ -15,7 +15,7 @@
 package com.google.maps.flutter.navigation
 
 class GoogleMapsNavigationInspectorHandler(
-  private val viewRegistry: GoogleMapsNavigationViewRegistry
+  private val viewRegistry: GoogleMapsViewRegistry
 ) : NavigationInspector {
   private fun manager(): GoogleMapsNavigationSessionManager {
     return GoogleMapsNavigationSessionManager.getInstance()
@@ -23,7 +23,7 @@ class GoogleMapsNavigationInspectorHandler(
 
   override fun isViewAttachedToSession(viewId: Long): Boolean {
     /// Is session exists, it's automatically attached to any existing view.
-    if (viewRegistry.getView(viewId.toInt()) != null) {
+    if (viewRegistry.getNavigationView(viewId.toInt()) != null) {
       return manager().isInitialized()
     }
     return false
