+++ android/src/main/kotlin/com/google/maps/flutter/navigation/GoogleMapsNavigationPlugin.kt
@@ -24,28 +24,28 @@ import io.flutter.embedding.engine.plugins.lifecycle.FlutterLifecycleAdapter
 
 /** GoogleMapsNavigationPlugin */
 class GoogleMapsNavigationPlugin : FlutterPlugin, ActivityAware {
-  private lateinit var viewRegistry: GoogleMapsNavigationViewRegistry
-  private lateinit var viewMessageHandler: GoogleMapsNavigationViewMessageHandler
+  private lateinit var viewRegistry: GoogleMapsViewRegistry
+  private lateinit var viewMessageHandler: GoogleMapsViewMessageHandler
   private lateinit var imageRegistryMessageHandler: GoogleMapsImageRegistryMessageHandler
-  private lateinit var navigationViewEventApi: NavigationViewEventApi
+  private lateinit var navigationViewEventApi: ViewEventApi
   private lateinit var _binding: FlutterPlugin.FlutterPluginBinding
   private lateinit var lifecycle: Lifecycle
   private lateinit var imageRegistry: ImageRegistry
 
   override fun onAttachedToEngine(binding: FlutterPlugin.FlutterPluginBinding) {
-    viewRegistry = GoogleMapsNavigationViewRegistry()
+    viewRegistry = GoogleMapsViewRegistry()
     imageRegistry = ImageRegistry()
-    viewMessageHandler = GoogleMapsNavigationViewMessageHandler(viewRegistry)
-    NavigationViewApi.setUp(binding.binaryMessenger, viewMessageHandler)
+    viewMessageHandler = GoogleMapsViewMessageHandler(viewRegistry)
+    MapViewApi.setUp(binding.binaryMessenger, viewMessageHandler)
     imageRegistryMessageHandler = GoogleMapsImageRegistryMessageHandler(imageRegistry)
     ImageRegistryApi.setUp(binding.binaryMessenger, imageRegistryMessageHandler)
-    navigationViewEventApi = NavigationViewEventApi(binding.binaryMessenger)
+    navigationViewEventApi = ViewEventApi(binding.binaryMessenger)
     _binding = binding
     binding.applicationContext.registerComponentCallbacks(viewRegistry)
   }
 
   override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {
-    NavigationViewApi.setUp(binding.binaryMessenger, null) // Cleanup
+    MapViewApi.setUp(binding.binaryMessenger, null) // Cleanup
     ImageRegistryApi.setUp(binding.binaryMessenger, null)
     GoogleMapsNavigationSessionManager.destroyInstance()
     binding.applicationContext.unregisterComponentCallbacks(viewRegistry)
@@ -53,7 +53,7 @@ class GoogleMapsNavigationPlugin : FlutterPlugin, ActivityAware {
 
   override fun onAttachedToActivity(binding: ActivityPluginBinding) {
     val factory =
-      GoogleMapsNavigationViewFactory(viewRegistry, navigationViewEventApi, imageRegistry)
+      GoogleMapsViewFactory(viewRegistry, navigationViewEventApi, imageRegistry)
     _binding.platformViewRegistry.registerViewFactory("google_navigation_flutter", factory)
     GoogleMapsNavigationSessionManager.createInstance(_binding.binaryMessenger)
     val inspectorHandler = GoogleMapsNavigationInspectorHandler(viewRegistry)
